generate_keyframes: Add arg "header", other minor changes
Add a bool for including the header information. This allows the generated keyframes to be used for x264's `--qpfile` without it throwing warnings.

Similarly, x264 may error out on `0 I -1` (observable with x264` 0.161.3027`, although it seems not to happen with `0.161.3018`). As it's also useless for timing, I opted to have it just not write a line for it at all.

Progress printing was also updated to update more frequently and replace the printed line as opposed to just outputting a bunch of them to reduce spam and makes it more obvious if it's freezing for some reason.

Other minor changes are removing an unneccesary comma in the docstrings and adding typehinting.